+++ src/main/java/org/jboss/qa/jenkins/test/executor/phase/runtimesetup/RuntimeSetup.java
@@ -37,5 +37,5 @@ public @interface RuntimeSetup {
 	String download() default "";
 
 	@Order
-	int order() default 0;
+	double order() default 0;
 }
